Update application version and revise back button placement
The version of the application has been updated from 0.6.23 to 0.6.24. Also, the back button in the simulation.py and emulation.py files was revised. Instead of being added to the main layout directly in simulation.py, it is now managed in the associated widget in emulation.py, improving code readability and organization.