Add new menu, main, and start screen files; update Python version
This commit introduces menu.py, main.py, and start_screen.py, which set up the main user interface. Python version in the project configurations is also updated from 3.9 to 3.12. The method 'show_hilfe' was made static, and the initialization order of 'menu_bar' and 'start_screen' in main.py was altered.